Move BaseAppWindow to lib
