+++ src/ui/apps/settings/settings.ipc.ts
@@ -2,7 +2,7 @@ import { generateUISchema } from '@/core/infrastructure/config/schema-utils';
 import { PROVIDER_MODELS, getProviderModelsWithDefaults } from '@core/infrastructure/config/config';
 import { ConfigurationManager } from '@core/infrastructure/config/configuration-manager';
 import { AppIpcModule, AppMainProcessInstances } from '@core/platform/ipc/types';
-import { BaseAppWindow } from '@ui/common/base-app-window';
+import { BaseAppWindow } from '@lib/base-app-window';
 import * as logger from '@utils/logger';
 import { IpcMain } from 'electron';
