+++ src/ui/apps/settings/settings.main.ts
@@ -1,7 +1,7 @@
 import { getWindowRegistry } from '@core/platform/windows/window-registry';
 import * as logger from '@utils/logger';
 import { BrowserWindow, Display } from 'electron';
-import { BaseAppWindow } from '@ui/common/base-app-window';
+import { BaseAppWindow } from '@lib/base-app-window';
 
 export class SettingsWindow extends BaseAppWindow {
     public window: BrowserWindow;
