+++ src/ui/platform/InputPill/inputpill.ipc.ts
@@ -1,6 +1,6 @@
 // src/apps/InputPill/input-pill.ipc.ts
 import { AppIpcModule, AppMainProcessInstances } from '@core/platform/ipc/types';
-import { BaseAppWindow } from '@ui/common/base-app-window';
+import { BaseAppWindow } from '@lib/base-app-window';
 import * as logger from '@utils/logger';
 import { IpcMain } from 'electron';
